Move ephemeral document store out of store module
